Factor out playing states into video-interface
