+++ test/integration/test-video-players-helper.js
@@ -132,7 +132,7 @@ export function runVideoPlayerIntegrationTests(
     afterEach(cleanUp);
   });
 
-  describe.configure().skipSaucelabs().run('Analytics Triggers', function() {
+  describe.configure().skipSauceLabs().run('Analytics Triggers', function() {
     this.timeout(TIMEOUT);
     let video;
